feat: get rid of chakra, vechain kit and tanstack
