+++ cross-app-connect/src/app/components/AccountChip.tsx
@@ -1,29 +1,13 @@
 'use client';
 
 import { useEffect, useState } from 'react';
-import {
-    Box,
-    HStack,
-    Icon,
-    IconButton,
-    Image,
-    Skeleton,
-    SkeletonCircle,
-    Stack,
-    Text,
-    Tooltip,
-    useClipboard,
-} from '@chakra-ui/react';
 import { LuCheck, LuCopy } from 'react-icons/lu';
 import { Address } from '@vechain/sdk-core';
-import type { ThorClient } from '@vechain/sdk-network';
-import {
-    useGetAvatarOfAddress,
-    useVechainDomain,
-} from '@vechain/vechain-kit';
-import { executeCallClause } from '@vechain/vechain-kit/utils';
 import { formatUnits, parseAbi } from 'viem';
+import { thor } from '../cross-app/_lib/thor';
+import { useAddressInfo } from '../cross-app/_lib/useAddressInfo';
 import type { TokenInfo } from '../cross-app/_lib/decoder';
+import styles from './AccountChip.module.css';
 
 const ERC20_BALANCE_ABI = parseAbi([
     'function balanceOf(address owner) view returns (uint256)',
@@ -31,7 +15,6 @@ const ERC20_BALANCE_ABI = parseAbi([
 
 type Props = {
     address: string;
-    thor: ThorClient | null;
     /**
      * Additional tokens (beyond native VET) to display the user's balance
      * for. Typically the tokens the current transaction touches.
@@ -65,13 +48,11 @@ function formatVET(raw: bigint): string {
  * its VeChain domain (if any) plus the avatar (custom for .vet domains,
  * Picasso identicon as fallback), a copy button, and the live VET balance.
  */
-export function AccountChip({ address, thor, relevantTokens }: Props) {
-    const { onCopy, hasCopied } = useClipboard(address);
+export function AccountChip({ address, relevantTokens }: Props) {
     const [balances, setBalances] = useState<LiveBalance[] | null>(null);
-    const { data: domainInfo, isPending: domainPending } =
-        useVechainDomain(address);
-    const { data: avatar, isPending: avatarPending } =
-        useGetAvatarOfAddress(address);
+    const [copied, setCopied] = useState(false);
+    const { domain, avatar, isLoading: addressInfoLoading } =
+        useAddressInfo(address);
     const balancesPending = balances === null;
 
     const tokenKey = (relevantTokens ?? [])
@@ -80,13 +61,10 @@ export function AccountChip({ address, thor, relevantTokens }: Props) {
         .join(',');
 
     useEffect(() => {
-        if (!thor || !address) return;
+        if (!address) return;
         let cancelled = false;
         (async () => {
             try {
-                // VET always shown first; then any tokens the transaction
-                // touches. Single Thor batch via Promise.all -- each is its
-                // own request but they fire in parallel.
                 const vetTask = thor.accounts
                     .getAccount(Address.of(address))
                     .then((acc) => ({
@@ -98,13 +76,11 @@ export function AccountChip({ address, thor, relevantTokens }: Props) {
                         (t) => t.address !== 'VET' && t.address !== 'vet',
                     )
                     .map(async (token) => {
-                        const res = await executeCallClause({
-                            thor,
-                            contractAddress: token.address,
-                            abi: ERC20_BALANCE_ABI,
-                            method: 'balanceOf' as const,
-                            args: [address as `0x${string}`],
-                        });
+                        const res = await thor.contracts
+                            .load(token.address, ERC20_BALANCE_ABI)
+                            .read.balanceOf(
+                                address as `0x${string}`,
+                            );
                         return {
                             token,
                             raw: BigInt(
@@ -124,146 +100,83 @@ export function AccountChip({ address, thor, relevantTokens }: Props) {
         return () => {
             cancelled = true;
         };
-    }, [thor, address, tokenKey]); // eslint-disable-line react-hooks/exhaustive-deps
+    }, [address, tokenKey]); // eslint-disable-line react-hooks/exhaustive-deps
 
-    const domain = domainInfo?.domain;
+    const handleCopy = async () => {
+        try {
+            await navigator.clipboard.writeText(address);
+            setCopied(true);
+            setTimeout(() => setCopied(false), 1500);
+        } catch {
+            // Clipboard API can fail in some browsers; silently ignore.
+        }
+    };
 
     return (
-        <HStack
-            spacing={3}
-            p={3}
-            rounded="md"
-            bg="card-elevated-bg"
-            borderWidth="1px"
-            borderColor="card-border"
-            align="center"
-        >
-            <SkeletonCircle
-                size="36px"
-                isLoaded={!avatarPending}
-                startColor="login-btn-hover-bg"
-                endColor="card-border"
-            >
-                {avatar ? (
-                    <Image
-                        src={avatar}
-                        alt=""
-                        boxSize="36px"
-                        rounded="full"
-                        draggable={false}
-                        fallback={
-                            <Box
-                                boxSize="36px"
-                                rounded="full"
-                                bg="login-btn-hover-bg"
-                            />
-                        }
-                    />
-                ) : (
-                    <Box
-                        boxSize="36px"
-                        rounded="full"
-                        bg="login-btn-hover-bg"
-                    />
-                )}
-            </SkeletonCircle>
-            <Stack spacing={0} flex={1} minW={0}>
-                <Text fontSize="xs" color="text-subtle">
-                    Your account
-                </Text>
-                <HStack spacing={2} align="center">
-                    <Skeleton
-                        isLoaded={!domainPending}
-                        startColor="login-btn-hover-bg"
-                        endColor="card-border"
-                        rounded="sm"
-                    >
-                        {domain ? (
-                            <Stack spacing={0}>
-                                <Text
-                                    fontWeight={600}
-                                    color="text-strong"
-                                    lineHeight="1.2"
-                                >
-                                    {domain}
-                                </Text>
-                                <Text
-                                    fontFamily="mono"
-                                    fontSize="xs"
-                                    color="text-subtle"
-                                    lineHeight="1.2"
-                                >
-                                    {truncate(address)}
-                                </Text>
-                            </Stack>
-                        ) : (
-                            <Text
-                                fontFamily="mono"
-                                fontSize="sm"
-                                fontWeight={500}
-                                color="text-strong"
-                            >
-                                {truncate(address)}
-                            </Text>
-                        )}
-                    </Skeleton>
-                    <Tooltip
-                        label={hasCopied ? 'Copied' : 'Copy address'}
-                        placement="top"
-                        hasArrow
-                        openDelay={150}
-                        fontSize="xs"
-                    >
-                        <IconButton
-                            aria-label="Copy address"
-                            icon={
-                                <Icon
-                                    as={hasCopied ? LuCheck : LuCopy}
-                                    boxSize="14px"
-                                />
-                            }
-                            size="xs"
-                            variant="ghost"
-                            h="22px"
-                            minW="22px"
-                            border="none"
-                            rounded="md"
-                            onClick={(e) => {
-                                e.stopPropagation();
-                                onCopy();
-                            }}
+        <div className={styles.chip}>
+            {addressInfoLoading ? (
+                <div className={`${styles.skeleton} ${styles.skeletonAvatar}`} />
+            ) : avatar ? (
+                <img
+                    src={avatar}
+                    alt=""
+                    className={styles.avatar}
+                    draggable={false}
+                />
+            ) : (
+                <div className={styles.avatarPlaceholder} />
+            )}
+            <div className={styles.identity}>
+                <p className={styles.label}>Your account</p>
+                <div className={styles.identityRow}>
+                    {addressInfoLoading ? (
+                        <div
+                            className={`${styles.skeleton} ${styles.skeletonDomain}`}
                         />
-                    </Tooltip>
-                </HStack>
-            </Stack>
-            <Stack spacing={0} align="flex-end">
-                <Text fontSize="xs" color="text-subtle">
-                    Balance
-                </Text>
+                    ) : domain ? (
+                        <div className={styles.domainStack}>
+                            <p className={styles.domain}>{domain}</p>
+                            <p className={styles.addressUnderDomain}>
+                                {truncate(address)}
+                            </p>
+                        </div>
+                    ) : (
+                        <p className={styles.addressOnly}>{truncate(address)}</p>
+                    )}
+                    <button
+                        type="button"
+                        className={styles.copyBtn}
+                        onClick={(e) => {
+                            e.stopPropagation();
+                            handleCopy();
+                        }}
+                        aria-label={copied ? 'Copied' : 'Copy address'}
+                        title={copied ? 'Copied' : 'Copy address'}
+                    >
+                        {copied ? (
+                            <LuCheck className={styles.copyIcon} />
+                        ) : (
+                            <LuCopy className={styles.copyIcon} />
+                        )}
+                    </button>
+                </div>
+            </div>
+            <div className={styles.balanceCol}>
+                <p className={styles.label}>Balance</p>
                 {balancesPending ? (
-                    <Skeleton
-                        height="14px"
-                        width="70px"
-                        rounded="sm"
-                        startColor="login-btn-hover-bg"
-                        endColor="card-border"
+                    <div
+                        className={`${styles.skeleton} ${styles.skeletonText}`}
                     />
                 ) : (
-                    <Stack spacing={0} align="flex-end">
+                    <div className={styles.balanceList}>
                         {balances!.map((b) => (
-                            <Text
-                                key={b.token.address}
-                                fontSize="sm"
-                                fontWeight={600}
-                                color="text-strong"
-                                fontFamily="mono"
-                            >
+                            <p key={b.token.address} className={styles.balanceRow}>
                                 {formatVET(b.raw)} {b.token.symbol}
-                            </Text>
+                            </p>
                         ))}
-                    </Stack>
+                    </div>
                 )}
-            </Stack>
-        </HStack>
+            </div>
+        </div>
     );
 }
