+++ cross-app-connect/src/app/components/AddressTag.tsx
@@ -1,28 +1,21 @@
 'use client';
 
-import { HStack, Icon, Image, Stack, Text, Tooltip } from '@chakra-ui/react';
 import { LuCircleCheck, LuTriangleAlert } from 'react-icons/lu';
-import {
-    useGetAvatarOfAddress,
-    useVechainDomain,
-    type AppConfig,
-} from '@vechain/vechain-kit';
 import { resolveContractLabel } from '../cross-app/_lib/contracts';
+import { useAddressInfo } from '../cross-app/_lib/useAddressInfo';
+import styles from './AddressTag.module.css';
 
 type Props = {
     address: string;
-    appConfig?: AppConfig;
     self?: string;
     /**
      * Distinguishes "contract this user is calling" (`'contract'` — default)
      * from "address receiving funds" (`'recipient'`).
      *
      * `'contract'` shows an "Unverified contract" warning when the address
-     * isn't in the kit's appConfig — that's the phishing-defence path.
-     *
-     * `'recipient'` is for token transfer destinations (a 2nd-leg argument
-     * to `transfer(address,uint256)` — typically just a wallet address, not
-     * a contract at all). Showing "Unverified contract" there would be
+     * isn't in the known-contracts registry — that's the phishing-defence
+     * path. `'recipient'` is for token transfer destinations (typically a
+     * wallet, not a contract); showing "Unverified contract" there would be
      * scary noise.
      */
     kind?: 'contract' | 'recipient';
@@ -37,125 +30,67 @@ function truncate(addr: string): string {
 
 export function AddressTag({
     address,
-    appConfig,
     self,
     kind = 'contract',
     avatarSize = 20,
 }: Props) {
-    // Hooks must be unconditional — React Query caches across instances so
-    // multiple AddressTags pointing at the same address share one fetch.
-    const { data: domainInfo } = useVechainDomain(address);
-    const { data: avatar } = useGetAvatarOfAddress(address);
-    const resolved = resolveContractLabel(address, appConfig, self);
+    const { domain, avatar } = useAddressInfo(address);
+    const resolved = resolveContractLabel(address, self);
 
-    // Verified VeChain-maintained contract: avatar (Picasso identicon)
-    // + friendly label + check. Avatar is here for visual consistency with
-    // unresolved addresses -- otherwise verified contracts feel like a
-    // different family.
     if (resolved) {
         return (
-            <HStack spacing={2} align="center">
+            <span className={styles.row}>
                 {avatar && (
-                    <Image
+                    <img
                         src={avatar}
                         alt=""
-                        boxSize={`${avatarSize}px`}
-                        rounded="full"
+                        width={avatarSize}
+                        height={avatarSize}
+                        className={styles.avatar}
                         draggable={false}
-                        fallback={
-                            <span
-                                style={{
-                                    width: avatarSize,
-                                    height: avatarSize,
-                                }}
-                            />
-                        }
                     />
                 )}
-                <Text fontWeight={500} color="text-strong">
-                    {resolved.label}
-                </Text>
+                <span className={styles.label}>{resolved.label}</span>
                 {resolved.verified && (
-                    <Tooltip
-                        label="Verified VeChain contract"
-                        placement="top"
-                        hasArrow
-                        openDelay={200}
-                        fontSize="xs"
-                    >
-                        <span style={{ display: 'inline-flex' }}>
-                            <Icon
-                                as={LuCircleCheck}
-                                color="green.500"
-                                boxSize="14px"
-                                aria-label="Verified"
-                            />
-                        </span>
-                    </Tooltip>
+                    <LuCircleCheck
+                        className={styles.iconVerified}
+                        aria-label="Verified"
+                        title="Verified VeChain contract"
+                    />
                 )}
-            </HStack>
+            </span>
         );
     }
 
-    // Unresolved: avatar + (domain ?? truncated address). Warn only when
-    // we're rendering the contract the user is *calling*, not a destination
-    // wallet for a token transfer.
-    const domain = domainInfo?.domain;
     return (
-        <HStack spacing={2} align="center" minW={0}>
+        <span className={styles.row}>
             {avatar && (
-                <Image
+                <img
                     src={avatar}
                     alt=""
-                    boxSize={`${avatarSize}px`}
-                    rounded="full"
+                    width={avatarSize}
+                    height={avatarSize}
+                    className={styles.avatar}
                     draggable={false}
-                    fallback={<span style={{ width: avatarSize, height: avatarSize }} />}
                 />
             )}
             {domain ? (
-                <Stack spacing={0} minW={0}>
-                    <Text
-                        fontWeight={500}
-                        color="text-strong"
-                        fontSize="sm"
-                        lineHeight="1.1"
-                        noOfLines={1}
-                    >
-                        {domain}
-                    </Text>
-                    <Text
-                        fontFamily="mono"
-                        fontSize="xs"
-                        color="text-subtle"
-                        lineHeight="1.1"
-                    >
+                <span className={styles.domainStack}>
+                    <span className={styles.domainName}>{domain}</span>
+                    <span className={styles.addressLine}>
                         {truncate(address)}
-                    </Text>
-                </Stack>
+                    </span>
+                </span>
             ) : (
-                <Text fontFamily="mono" fontSize="sm" color="text-muted">
-                    {truncate(address)}
-                </Text>
+                <span className={styles.addressOnly}>{truncate(address)}</span>
             )}
             {kind === 'contract' && (
-                <Tooltip
-                    label="Unverified contract — make sure you trust it before continuing"
-                    placement="top"
-                    hasArrow
-                    openDelay={200}
-                    fontSize="xs"
-                >
-                    <span style={{ display: 'inline-flex' }}>
-                        <Icon
-                            as={LuTriangleAlert}
-                            color="orange.400"
-                            boxSize="14px"
-                            aria-label="Unverified"
-                        />
-                    </span>
-                </Tooltip>
+                <LuTriangleAlert
+                    className={styles.iconWarn}
+                    aria-label="Unverified"
+                    title="Unverified contract — make sure you trust it before continuing"
+                />
             )}
-        </HStack>
+        </span>
     );
 }
