+++ cross-app-connect/src/app/components/RequesterChip.tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import { useState } from 'react';
-import { HStack, Icon, Image, Text, Tooltip } from '@chakra-ui/react';
 import {
     LuCircleCheck,
     LuGlobe,
@@ -9,6 +8,7 @@ import {
     LuTriangleAlert,
 } from 'react-icons/lu';
 import { lookupAppByUrl } from '../cross-app/_lib/app-hub';
+import styles from './RequesterChip.module.css';
 
 type Props = {
     url: string;
@@ -27,11 +27,7 @@ export function RequesterChip({ url }: Props) {
     const [iconBroken, setIconBroken] = useState(false);
     const parsed = safeParseUrl(url);
     if (!parsed) {
-        return (
-            <Text fontSize="sm" color="text-muted">
-                {url}
-            </Text>
-        );
+        return <span className={styles.fallback}>{url}</span>;
     }
 
     const isSecure = parsed.protocol === 'https:';
@@ -42,81 +38,43 @@ export function RequesterChip({ url }: Props) {
     const faviconSrc = `https://www.google.com/s2/favicons?domain=${parsed.hostname}&sz=64`;
     const appHubEntry = lookupAppByUrl(url);
     const verified = Boolean(appHubEntry);
+    const SecurityIcon = isSecure ? LuLockKeyhole : LuGlobe;
+    const VerifiedIcon = verified ? LuCircleCheck : LuTriangleAlert;
 
     return (
-        <HStack
-            spacing={2}
-            px={2}
-            py="3px"
-            rounded="full"
-            bg="card-bg"
-            borderWidth="1px"
-            borderColor={verified ? 'green.500' : 'login-btn-border'}
-            display="inline-flex"
-            maxW="full"
+        <span
+            className={`${styles.chip} ${verified ? styles.chipVerified : ''}`}
         >
-            <Icon
-                as={isSecure ? LuLockKeyhole : LuGlobe}
-                boxSize="14px"
-                color={isSecure ? 'green.500' : 'text-subtle'}
+            <SecurityIcon
+                className={`${styles.icon} ${isSecure ? styles.iconSecure : styles.iconInsecure}`}
                 aria-label={
                     isSecure ? 'Secure (HTTPS)' : 'Not encrypted (HTTP)'
                 }
             />
             {!iconBroken && (
-                <Image
+                <img
                     src={faviconSrc}
                     alt=""
-                    boxSize="16px"
-                    rounded="sm"
+                    className={styles.favicon}
                     onError={() => setIconBroken(true)}
                     draggable={false}
                 />
             )}
-            <Text
-                fontSize="sm"
-                fontWeight={500}
-                color="text-strong"
-                noOfLines={1}
-            >
+            <span className={styles.label}>
                 {verified ? appHubEntry!.name : display}
-            </Text>
-            {verified ? (
-                <Tooltip
-                    label="Listed in the VeChain App Hub"
-                    placement="top"
-                    hasArrow
-                    openDelay={150}
-                    fontSize="xs"
-                >
-                    <span style={{ display: 'inline-flex' }}>
-                        <Icon
-                            as={LuCircleCheck}
-                            boxSize="14px"
-                            color="green.500"
-                            aria-label="Verified VeChain app"
-                        />
-                    </span>
-                </Tooltip>
-            ) : (
-                <Tooltip
-                    label="Not listed in the VeChain App Hub — proceed only if you recognize this site"
-                    placement="top"
-                    hasArrow
-                    openDelay={150}
-                    fontSize="xs"
-                >
-                    <span style={{ display: 'inline-flex' }}>
-                        <Icon
-                            as={LuTriangleAlert}
-                            boxSize="14px"
-                            color="orange.400"
-                            aria-label="Unverified app"
-                        />
-                    </span>
-                </Tooltip>
-            )}
-        </HStack>
+            </span>
+            <VerifiedIcon
+                className={`${styles.icon} ${verified ? styles.iconVerified : styles.iconUnverified}`}
+                aria-label={
+                    verified ? 'Verified VeChain app' : 'Unverified app'
+                }
+                title={
+                    verified
+                        ? 'Listed in the VeChain App Hub'
+                        : 'Not listed in the VeChain App Hub — proceed only if you recognize this site'
+                }
+            />
+        </span>
     );
 }
