+++ cross-app-connect/src/app/components/VechainHeader.tsx
@@ -1,8 +1,6 @@
-'use client';
-
-import { Box, HStack, Icon, Image, Stack, Text } from '@chakra-ui/react';
 import type { IconType } from 'react-icons';
 import { RequesterChip } from './RequesterChip';
+import styles from './VechainHeader.module.css';
 
 type Props = {
     title?: string;
@@ -13,68 +11,52 @@ type Props = {
      */
     titleIcon?: IconType;
     /**
-     * Color token for the title icon. Defaults to 'accent'. The transact
-     * screen passes 'orange.400' / 'red.400' on cautioned / dangerous
+     * Color for the title icon. Defaults to the accent token. The transact
+     * screen passes `var(--warn)` / `var(--danger)` on cautioned / dangerous
      * transactions so the icon swaps in tandem with the verb.
      */
     titleIconColor?: string;
     subtitle?: string;
     /**
      * Requester dApp's callbackUrl. When provided, renders a chip with the
-     * site's favicon + hostname under the title to identify who's asking
-     * to connect.
+     * site's favicon + hostname under the title to identify who's asking.
      */
     requesterUrl?: string;
 };
 
 export function VechainHeader({
     title = 'Log in to your wallet',
-    titleIcon,
-    titleIconColor = 'accent',
+    titleIcon: TitleIcon,
+    titleIconColor = 'var(--accent)',
     subtitle,
     requesterUrl,
 }: Props) {
-    // Render BOTH logomarks and toggle via Chakra's _light / _dark CSS
-    // pseudo selectors. Avoids the React-state flicker we'd get if we
-    // picked the src from useColorMode() (SSR default vs client value).
     return (
-        <Stack spacing={4} align="center" pt={6} pb={2}>
-            <Box boxSize="48px" display="flex" alignItems="center">
-                <Image
+        <header className={styles.header}>
+            <div className={styles.logoBox}>
+                <img
                     src="/brand/vechain-logomark-light.png"
                     alt="VeChain"
-                    boxSize="48px"
+                    className={styles.logo}
                     draggable={false}
                 />
-            </Box>
-            <Stack spacing={2} align="center" textAlign="center" maxW="sm">
-                <HStack spacing={2} align="center">
-                    {titleIcon && (
-                        <Icon
-                            as={titleIcon}
-                            color={titleIconColor}
-                            boxSize="18px"
+            </div>
+            <div className={styles.titleStack}>
+                <div className={styles.titleRow}>
+                    {TitleIcon && (
+                        <TitleIcon
+                            className={styles.titleIcon}
+                            style={{ color: titleIconColor }}
                             aria-hidden
                         />
                     )}
-                    <Text
-                        fontFamily="heading"
-                        fontWeight={600}
-                        fontSize="2xl"
-                        color="text-strong"
-                    >
-                        {title}
-                    </Text>
-                </HStack>
+                    <h1 className={styles.title}>{title}</h1>
+                </div>
 
-                {subtitle && (
-                    <Text fontSize="lg" color="text-muted">
-                        {subtitle}
-                    </Text>
-                )}
+                {subtitle && <p className={styles.subtitle}>{subtitle}</p>}
 
                 {requesterUrl && <RequesterChip url={requesterUrl} />}
-            </Stack>
-        </Stack>
+            </div>
+        </header>
     );
 }
