+++ cross-app-connect/src/app/cross-app/transact/transact.module.css
@@ -1,0 +1,321 @@
+.shell {
+    max-width: 28rem;
+    margin: 0 auto;
+    padding: 32px 16px;
+    display: flex;
+    flex-direction: column;
+    gap: 24px;
+}
+
+.card {
+    background: var(--bg-card);
+    border-radius: var(--radius-md);
+    padding: 16px;
+}
+
+.cardBody {
+    display: flex;
+    flex-direction: column;
+    gap: 16px;
+}
+
+.center {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding: 40px 0;
+}
+
+.spinner {
+    width: 32px;
+    height: 32px;
+    border: 3px solid var(--border-button);
+    border-top-color: var(--accent);
+    border-radius: 50%;
+    animation: spin 0.8s linear infinite;
+}
+
+@keyframes spin {
+    to { transform: rotate(360deg); }
+}
+
+.fallbackText {
+    color: var(--text-muted);
+    text-align: center;
+    font-size: 14px;
+    line-height: 1.4;
+    margin: 0;
+}
+
+.alert {
+    display: flex;
+    align-items: flex-start;
+    gap: 8px;
+    padding: 12px;
+    border-radius: var(--radius-md);
+    border-left: 3px solid;
+    font-size: 12px;
+    line-height: 1.3;
+}
+
+.alertError {
+    background: rgba(239, 68, 68, 0.08);
+    border-left-color: var(--danger);
+    color: var(--text-strong);
+}
+
+.alertWarn {
+    background: rgba(245, 158, 11, 0.08);
+    border-left-color: var(--warn);
+    color: var(--text-strong);
+}
+
+.btnBrand {
+    background: var(--primary-btn-bg);
+    color: var(--primary-btn-color);
+    border: none;
+    border-radius: var(--radius-full);
+    height: 48px;
+    padding: 0 16px;
+    font-weight: 500;
+    font-size: 15px;
+    width: 100%;
+    transition: opacity 0.2s;
+}
+
+.btnBrand:hover:not(:disabled) {
+    opacity: 0.85;
+}
+
+.btnBrand:disabled {
+    opacity: 0.5;
+}
+
+.btnGhost {
+    background: transparent;
+    color: var(--text-muted);
+    border: 1px solid var(--border-button);
+    border-radius: var(--radius-full);
+    height: 48px;
+    padding: 0 24px;
+    font-weight: 500;
+    width: 100%;
+    transition: all 0.2s;
+}
+
+.btnGhost:hover:not(:disabled) {
+    background: var(--login-btn-hover-bg);
+    color: var(--text-strong);
+}
+
+.btnGhost:disabled {
+    opacity: 0.4;
+}
+
+.linkBtn {
+    background: transparent;
+    border: none;
+    padding: 0;
+    color: var(--text-muted);
+    font-size: 14px;
+    font-weight: 500;
+    display: inline-flex;
+    align-items: center;
+    gap: 4px;
+}
+
+.linkBtn:hover {
+    color: var(--text-strong);
+    text-decoration: underline;
+}
+
+.inspectRow {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 4px;
+    padding-top: 4px;
+}
+
+.muted {
+    font-size: 12px;
+    color: var(--text-subtle);
+    margin: 0;
+}
+
+.actionList {
+    display: flex;
+    flex-direction: column;
+    gap: 12px;
+}
+
+.actionRow {
+    display: flex;
+    flex-direction: column;
+    padding-left: 12px;
+    border-left: 2px solid var(--border-button);
+}
+
+.actionRowWarn {
+    border-left-color: var(--warn);
+}
+
+.actionTitle {
+    font-weight: 600;
+    color: var(--text-strong);
+    margin: 0;
+}
+
+.actionDetail {
+    display: flex;
+    align-items: center;
+    gap: 6px;
+    font-size: 14px;
+    color: var(--text-muted);
+    margin: 0;
+}
+
+.skeletonStack {
+    display: flex;
+    flex-direction: column;
+    gap: 4px;
+    padding-left: 12px;
+    border-left: 2px solid var(--border-button);
+}
+
+.skeletonBar {
+    background: var(--login-btn-hover-bg);
+    border-radius: var(--radius-sm);
+    animation: pulse 1.5s ease-in-out infinite;
+}
+
+.skeletonBarLg {
+    height: 16px;
+    width: 55%;
+}
+
+.skeletonBarSm {
+    height: 12px;
+    width: 35%;
+}
+
+@keyframes pulse {
+    0%, 100% { opacity: 1; }
+    50% { opacity: 0.55; }
+}
+
+.subPanel {
+    padding: 12px;
+    border-radius: var(--radius-md);
+    background: var(--bg-card-elevated);
+    border: 1px solid var(--border-default);
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+}
+
+.subPanelLabel {
+    font-size: 12px;
+    color: var(--text-subtle);
+    text-transform: uppercase;
+    letter-spacing: 0.05em;
+    margin: 0 0 8px;
+}
+
+.messageBody {
+    font-size: 14px;
+    color: var(--text-strong);
+    white-space: pre-wrap;
+    word-break: break-word;
+    margin: 0;
+}
+
+.typedHead {
+    font-size: 12px;
+    color: var(--text-subtle);
+    text-transform: uppercase;
+    letter-spacing: 0.05em;
+    margin: 0;
+}
+
+.typedFrom {
+    font-size: 12px;
+    color: var(--text-muted);
+    margin: 0;
+}
+
+.code {
+    font-family: ui-monospace, 'SF Mono', Menlo, Consolas, monospace;
+    font-size: 12px;
+    color: var(--text-strong);
+    white-space: pre-wrap;
+    word-break: break-word;
+    background: transparent;
+    margin: 0;
+}
+
+.codeBlock {
+    font-family: ui-monospace, 'SF Mono', Menlo, Consolas, monospace;
+    font-size: 12px;
+    color: var(--text-muted);
+    white-space: pre-wrap;
+    word-break: break-all;
+    padding: 12px;
+    border-radius: var(--radius-md);
+    background: var(--bg-card-elevated);
+    border: 1px solid var(--border-default);
+    margin: 0;
+}
+
+.detailsList {
+    display: flex;
+    flex-direction: column;
+    gap: 12px;
+    margin-top: 8px;
+}
+
+.detailRow {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.detailLabel {
+    font-size: 12px;
+    color: var(--text-subtle);
+    margin: 0;
+}
+
+.detailValue {
+    font-family: ui-monospace, 'SF Mono', Menlo, Consolas, monospace;
+    font-size: 12px;
+    color: var(--text-muted);
+    margin: 0;
+}
+
+.clauseStack {
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+}
+
+.clauseHeader {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.collapse {
+    overflow: hidden;
+    transition: max-height 0.2s ease, opacity 0.2s ease;
+}
+
+.collapseOpen {
+    max-height: 2000px;
+    opacity: 1;
+}
+
+.collapseClosed {
+    max-height: 0;
+    opacity: 0;
+}
