+++ cross-app-connect/src/app/globals.css
@@ -1,7 +1,104 @@
+/* Design tokens mirror the defaults used by `@vechain/vechain-kit` so the
+   whitelabel host feels native next to consumer dApps. Source of truth was
+   `theme/brand.ts` before this migration; tokens are now consumed directly
+   as CSS variables. */
+
+:root {
+    /* surfaces */
+    --bg-page: #ffffff;
+    --bg-card: #f5f5f5;
+    --bg-card-elevated: #ffffff;
+    --border-default: transparent;
+    --border-button: #ebebeb;
+    --border-hover: #d0d0d0;
+
+    /* text */
+    --text-strong: #2e2e2e;
+    --text-muted: #4d4d4d;
+    --text-subtle: #718096;
+
+    /* buttons */
+    --login-btn-bg: #ffffff;
+    --login-btn-color: #1a1a1a;
+    --login-btn-hover-bg: #f0f0f0;
+    --primary-btn-bg: #272a2e;
+    --primary-btn-color: #ffffff;
+
+    /* accents */
+    --accent: #3b82f6;
+    --chip-bg: rgba(114, 102, 255, 0.12);
+    --chip-text: #5b50cc;
+
+    /* status */
+    --danger: #ef4444;
+    --warn: #f59e0b;
+    --success: #10b981;
+
+    /* typography */
+    --font-heading: "Satoshi", "Inter", system-ui, -apple-system, sans-serif;
+    --font-body: "Inter", system-ui, -apple-system, sans-serif;
+
+    /* radii */
+    --radius-sm: 8px;
+    --radius-md: 12px;
+    --radius-lg: 16px;
+    --radius-xl: 24px;
+    --radius-full: 9999px;
+}
+
+[data-color-mode='dark'] {
+    --bg-page: #151515;
+    --bg-card: rgba(255, 255, 255, 0.04);
+    --bg-card-elevated: #2a2a2a;
+    --border-default: rgba(255, 255, 255, 0.1);
+    --border-button: rgba(255, 255, 255, 0.1);
+    --border-hover: rgba(255, 255, 255, 0.2);
+
+    --text-strong: rgb(223, 223, 221);
+    --text-muted: rgba(223, 223, 221, 0.6);
+    --text-subtle: rgba(223, 223, 221, 0.4);
+
+    --login-btn-bg: transparent;
+    --login-btn-color: #ffffff;
+    --login-btn-hover-bg: rgba(255, 255, 255, 0.05);
+    --primary-btn-bg: #ffffff;
+    --primary-btn-color: rgba(0, 0, 0, 0.9);
+
+    --accent: #60a5fa;
+    --chip-bg: rgba(114, 102, 255, 0.2);
+    --chip-text: #b9b0ff;
+}
+
 html,
 body {
     margin: 0;
     min-height: 100vh;
     -webkit-font-smoothing: antialiased;
     -moz-osx-font-smoothing: grayscale;
+}
+
+body {
+    background: var(--bg-page);
+    color: var(--text-strong);
+    font-family: var(--font-body);
+}
+
+*,
+*::before,
+*::after {
+    box-sizing: border-box;
+}
+
+button {
+    font-family: inherit;
+    cursor: pointer;
+}
+
+button:disabled {
+    cursor: not-allowed;
+}
+
+a {
+    color: inherit;
+    text-decoration: none;
 }
