--- cross-app-connect/src/app/theme/index.ts
@@ -1,2 +1,0 @@
-export { vechainTheme } from './theme';
-export { tokens } from './brand';
